requirements-bot.txt
discord.py>=2.3.0
gspread>=5.12.0
oauth2client>=4.1.3
